+++ packages/core/src/awsService/apprunner/wizards/apprunnerCreateServiceWizard.ts
@@ -15,7 +15,7 @@ import { AppRunnerCodeRepositoryWizard } from './codeRepositoryWizard'
 import { GitExtension } from '../../../shared/extensions/git'
 import { makeDeploymentButton } from './deploymentButton'
 import { createExitPrompter } from '../../../shared/ui/common/exitPrompter'
-import { DefaultIamClient } from '../../../shared/clients/iamClient'
+import { DefaultIamClient } from '../../../shared/clients/iam'
 import { DefaultEcrClient } from '../../../shared/clients/ecrClient'
 import { DefaultAppRunnerClient } from '../../../shared/clients/apprunnerClient'
 import { getAppRunnerCreateServiceDocUrl } from '../../../shared/extensionUtilities'
