+++ packages/core/src/awsService/apprunner/wizards/imageRepositoryWizard.ts
@@ -17,7 +17,7 @@ import * as nls from 'vscode-nls'
 import { WizardForm } from '../../../shared/wizards/wizardForm'
 import { createVariablesPrompter } from '../../../shared/ui/common/variablesPrompter'
 import { makeDeploymentButton } from './deploymentButton'
-import { IamClient } from '../../../shared/clients/iamClient'
+import { IamClient } from '../../../shared/clients/iam'
 import { createRolePrompter } from '../../../shared/ui/common/roles'
 import { getLogger } from '../../../shared/logger/logger'
 import { getAppRunnerCreateServiceDocUrl, isCloud9 } from '../../../shared/extensionUtilities'
