+++ packages/core/src/awsService/ecs/util.ts
@@ -9,7 +9,7 @@ import * as nls from 'vscode-nls'
 const localize = nls.loadMessageBundle()
 
 import { EcsClient } from '../../shared/clients/ecsClient'
-import { DefaultIamClient, IamClient } from '../../shared/clients/iamClient'
+import { DefaultIamClient, IamClient } from '../../shared/clients/iam'
 import { ToolkitError } from '../../shared/errors'
 import { isCloud9 } from '../../shared/extensionUtilities'
 import { getOrInstallCli } from '../../shared/utilities/cliUtils'
