+++ packages/core/src/test/awsService/apprunner/wizards/createServiceWizard.test.ts
@@ -7,7 +7,7 @@ import { createWizardTester, WizardTester } from '../../../shared/wizards/wizard
 import { AppRunner } from 'aws-sdk'
 import { CreateAppRunnerServiceWizard } from '../../../../awsService/apprunner/wizards/apprunnerCreateServiceWizard'
 import { stub } from '../../../utilities/stubber'
-import { DefaultIamClient } from '../../../../shared/clients/iamClient'
+import { DefaultIamClient } from '../../../../shared/clients/iam'
 import { DefaultEcrClient } from '../../../../shared/clients/ecrClient'
 import { DefaultAppRunnerClient } from '../../../../shared/clients/apprunnerClient'
