+++ packages/core/src/test/awsService/ec2/model.test.ts
@@ -12,7 +12,7 @@ import { Ec2Selection } from '../../../awsService/ec2/prompter'
 import { ToolkitError } from '../../../shared/errors'
 import { IAM } from 'aws-sdk'
 import { SshKeyPair } from '../../../awsService/ec2/sshKeyPair'
-import { DefaultIamClient } from '../../../shared/clients/iamClient'
+import { DefaultIamClient } from '../../../shared/clients/iam'
 import { assertNoTelemetryMatch, createTestWorkspaceFolder } from '../../testUtil'
 import { fs } from '../../../shared'
 import path from 'path'
