+++ packages/core/src/test/awsService/ecs/util.test.ts
@@ -6,7 +6,7 @@
 import assert from 'assert'
 import { stub } from '../../utilities/stubber'
 import { checkPermissionsForSsm } from '../../../awsService/ecs/util'
-import { DefaultIamClient } from '../../../shared/clients/iamClient'
+import { DefaultIamClient } from '../../../shared/clients/iam'
 
 describe('checkPermissionsForSsm', function () {
     const getClient = () => stub(DefaultIamClient, { regionCode: '' })
