+++ packages/core/src/test/shared/ui/prompters/roles.test.ts
@@ -7,7 +7,7 @@ import * as vscode from 'vscode'
 import * as sinon from 'sinon'
 import assert from 'assert'
 import { IAM } from 'aws-sdk'
-import { DefaultIamClient } from '../../../../shared/clients/iamClient'
+import { DefaultIamClient } from '../../../../shared/clients/iam'
 import { createQuickPickPrompterTester, QuickPickPrompterTester } from '../testUtils'
 import { createRolePrompter } from '../../../../shared/ui/common/roles'
 import { toCollection } from '../../../../shared/utilities/asyncCollection'
